added lambda and remove min Func

+++ PredictB0B1.cpp
@@ -2,8 +2,8 @@
 
 PredictB0B1::PredictB0B1(
     double _error,
-    int _b0,
-    int _b1
+    double _b0,
+    double _b1
 )
 :   error(_error),
     b0(_b0),
