+++ PredictB0B1.h
@@ -2,11 +2,11 @@ class PredictB0B1{
     public:
         PredictB0B1(
             double _error,
-            int _b0,
-            int _b1
+            double _b0,
+            double _b1
         );
 
         double error;
-        int b0;
-        int b1;
+        double b0;
+        double b1;
 };
